pyproject.toml
[project]
name = "agentic-ai-part-1"
version = "0.1.0"
description = "Code repository for Agentic AI training program"
authors = [
    { name = "Simanta Sarma", email = "ximanta@cc.cc" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "jupyter>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
    "pytest-cov>=4.0.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
